requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1
openai==1.1.1
anthropic==0.8.1
gunicorn==21.2.0
gevent==23.9.1
flask-cors==4.0.0
python-dateutil==2.8.2
wtforms==3.0.1 